+++ tests/run-tests.ps1
@@ -1,0 +1,5 @@
+﻿$moduleName = "PSIASTAND"
+$here = Split-Path -Parent $MyInvocation.MyCommand.Path
+Import-Module "$here\..\$($moduleName)"
+Invoke-Pester
+Remove-Module $moduleName