+++ tests/unit/sqlite/main/Invoke-SQLiteBulkCopy.tests.ps1
@@ -12,7 +12,7 @@ $moduleName = "PSIASTAND"
 
 $PSVersion = $PSVersionTable.PSVersion.Major
 
-Import-Module "$here\..\..\..\..\$($moduleName)"
+#Import-Module "$here\..\..\..\..\$($moduleName)"
 
 $SQLiteFile = "$here\Working.SQLite"
 Remove-Item $SQLiteFile  -force -ErrorAction SilentlyContinue
@@ -67,5 +67,5 @@ Describe "Invoke-SQLiteBulkCopy (Requires Out-DataTable) PS$PSVersion" {
     }
 }
 
-Remove-Module $moduleName
+#Remove-Module $moduleName
 Remove-Item $SQLiteFile -force -ErrorAction SilentlyContinue