+++ tests/unit/sqlite/main/New-SQLiteConnection.tests.ps1
@@ -12,7 +12,7 @@ $moduleName = "PSIASTAND"
 
 $PSVersion = $PSVersionTable.PSVersion.Major
 
-Import-Module "$here\..\..\..\..\$($moduleName)"
+#Import-Module "$here\..\..\..\..\$($moduleName)"
 
 $SQLiteFile = "$here\Working.SQLite"
 Remove-Item $SQLiteFile  -force -ErrorAction SilentlyContinue
@@ -33,5 +33,5 @@ Describe "New-SQLiteConnection PS$PSVersion" {
     }
 }
 
-Remove-Module $moduleName
+#Remove-Module $moduleName
 Remove-Item $SQLiteFile -force -ErrorAction SilentlyContinue