+++ tests/unit/support/Get-Timediff.Tests.ps1
@@ -4,7 +4,7 @@ $sut = (Split-Path -Leaf $MyInvocation.MyCommand.Path) -replace '\.Tests\.', '.'
 
 $moduleName = "PSIASTAND"
 
-Import-Module "$here\..\..\..\$($moduleName)"
+#Import-Module "$here\..\..\..\$($moduleName)"
 #InModuleScope nessusOpenPorts {
     Describe "Get-Timediff" {
         It "should return 1 hour difference" {
@@ -28,4 +28,4 @@ Import-Module "$here\..\..\..\$($moduleName)"
         }
     }
 #}
-Remove-Module $moduleName
+#Remove-Module $moduleName