+++ CHANGELOG.md
@@ -1,7 +1,13 @@
-﻿## 1.0.1 (Feb, 9, 2016)
+﻿## 1.0.1 (Feb, 10, 2016)
     - Implemented excel unit tests
     - Implemented SQL unit tests
     - Implemented Import-XML function and tests
+    - Removed the module loads and unloads from tests scripts
+    - create run-tests script to load module run pester and unload module
+    - Updated import-ckl to handle both version 1 and version 2 CKL files
+    - Created unit tests for Import-CKL
+    - Created ConvertTo-CKL and implemented unit tests
+    - Created Export-CKL end to end script and e2e tests
 
 ## 1.0.0 (Feb, 8, 2016)
     - Initial Release